requirements.txt
elevenlabs==0.2.21
openai==0.27.8
python-dotenv==1.0.0
streamlit==1.23.1
